% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combi.R
\name{combi}
\alias{combi}
\title{combi: compute the number of all possible subsets of size k from an n-size set (= k-combination)}
\usage{
combi(n, k)
}
\arguments{
\item{n}{is the number of elements in the set}

\item{k}{is the number of elements in the subset}
}
\description{
combi: compute the number of all possible subsets of size k from an n-size set (= k-combination)
}
\examples{
combi(20,5)
}
\keyword{combination,}
\keyword{combinatorics}

