% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectedJaccard.R
\name{expectedJaccard}
\alias{expectedJaccard}
\title{expectedJaccard: expected jaccard similarity between two subsets S and T}
\usage{
expectedJaccard(m, n)
}
\arguments{
\item{m}{is the size of subsets S and T}

\item{n}{is the population size (from which elements in S and T are drawn) - only small sizes!}
}
\description{
expectedJaccard: expected jaccard similarity between two subsets S and T
}
\examples{
expectedJaccard(5,20)
}
\keyword{expected}
\keyword{jaccard,}
\keyword{similarity,}

