% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectedJaccardApprox.R
\name{expectedJaccardApprox}
\alias{expectedJaccardApprox}
\title{expectedJaccardApprox: approximation of expected jaccard similarity between two subsets S and T}
\usage{
expectedJaccardApprox(m, n)
}
\arguments{
\item{m}{is the size of subsets S and T}

\item{n}{is the population size (from which elements in S and T are drawn)}
}
\description{
expectedJaccardApprox: approximation of expected jaccard similarity between two subsets S and T
}
\examples{
expectedJaccardApprox(10,1200)
}
\keyword{approximation}
\keyword{expected,}
\keyword{jaccard,}
\keyword{similarity,}

