% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frameToMat.R
\name{frameToMat}
\alias{frameToMat}
\title{frameToMat: applies a function (returning a float) to every pair of columns in a dataframe and outputs a matrix}
\usage{
frameToMat(x, FUN = afunc)
}
\arguments{
\item{x}{your dataframe}

\item{afunc}{(any) function that operates on two vectors and returns a float}
}
\description{
frameToMat: applies a function (returning a float) to every pair of columns in a dataframe and outputs a matrix
}
\examples{
mymatrix <- frameToMat(mydataframe,computeDistance)
}
\keyword{apply,}
\keyword{dataframe,}
\keyword{matrix,}
\keyword{similarity}

