% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringJaccard.R
\name{stringJaccard}
\alias{stringJaccard}
\title{stringJaccard: computes jaccard similarity between two string vectors}
\usage{
stringJaccard(x, y)
}
\arguments{
\item{x}{a string vector}

\item{y}{is another string vector}
}
\description{
stringJaccard: computes jaccard similarity between two string vectors
}
\examples{
jaccardSim <- stringJaccard(stringvector1, stringvector2)
}
\keyword{jaccard,}
\keyword{similarity,}
\keyword{string}

